refactor UI, Application
this breaks sen hard; will be finished with commands refactorization
