+++ sen/tui/init.py
@@ -1,171 +1,31 @@
+"""
+Application specific code.
+"""
+
 import logging
-import threading
-from concurrent.futures.thread import ThreadPoolExecutor
 
 from sen.exceptions import NotifyError
-from sen.tui.commands import search, filter
-from sen.tui.statusbar import UIFrameWidget
+from sen.tui.ui import get_app_in_loop
 from sen.tui.buffer import LogsBuffer, MainListBuffer, InspectBuffer, HelpBuffer, ImageInfoBuffer, TreeBuffer, \
     ContainerInfoBuffer
 from sen.tui.constants import PALLETE
 from sen.docker_backend import DockerBackend, DockerImage, DockerContainer
 
-import urwid
-
-from sen.util import log_traceback, OrderedSet
+from sen.util import log_traceback
 
 logger = logging.getLogger(__name__)
 
 
-class UI(urwid.MainLoop):
+class Application:
     def __init__(self):
         self.d = DockerBackend()
 
-        # root widget
-        self.mainframe = UIFrameWidget(self, urwid.SolidFill())
-        self.buffers = []
-        self.buffer_movement_history = OrderedSet()
-
-        self.refresh_lock = threading.Lock()
-
-        # worker for long-running tasks - requests
-        self.worker = ThreadPoolExecutor(max_workers=4)
-        # worker for quick ui operations
-        self.ui_worker = ThreadPoolExecutor(max_workers=2)
-
-        root_widget = urwid.AttrMap(self.mainframe, "root")
-        self.main_list_buffer = None  # singleton
-
-        screen = urwid.raw_display.Screen()
-        screen.set_terminal_properties(256)
-        screen.register_palette(PALLETE)
-
-        super().__init__(root_widget, screen=screen, event_loop=urwid.AsyncioEventLoop())
-        self.handle_mouse = False
-        self.current_buffer = None
-
-    def run_in_background(self, task, *args, **kwargs):
-        logger.info("running task %r(%s, %s) in background", task, args, kwargs)
-        self.worker.submit(task, *args, **kwargs)
-
-    def run_quickly_in_bacakground(self, task, *args, **kwargs):
-        logger.info("running a quick task %r(%s, %s) in background", task, args, kwargs)
-        self.ui_worker.submit(task, *args, **kwargs)
-
-    def entering_idle(self):
-        with self.refresh_lock:
-            return super().entering_idle()
-
-    def refresh(self):
-        logger.debug("refresh user interface")
-        try:
-            with self.refresh_lock:
-                self.draw_screen()
-        except AssertionError:
-            logger.warning("application is not running")
-            pass
-
-    def _set_main_widget(self, widget, redraw):
-        """
-        add provided widget to widget list and display it
-
-        :param widget:
-        :return:
-        """
-        self.mainframe.set_body(widget)
-        self.reload_footer()
-        if redraw:
-            logger.debug("redraw main widget")
-            self.refresh()
-
-    def display_buffer(self, buffer, redraw=True):
-        """
-        display provided buffer
-
-        :param buffer: Buffer
-        :return:
-        """
-        self.buffer_movement_history.append(buffer)
-        logger.debug("movement history: %s", self.buffer_movement_history)
-        self.current_buffer = buffer
-        self._set_main_widget(buffer.widget, redraw=redraw)
-
-    def add_and_display_buffer(self, buffer, redraw=True):
-        """
-        add provided buffer to buffer list and display it
-
-        :param buffer:
-        :return:
-        """
-        if buffer not in self.buffers:
-            logger.debug("adding new buffer {!r}".format(buffer))
-            self.buffers.append(buffer)
-        self.display_buffer(buffer, redraw=redraw)
-
-    def pick_and_display_buffer(self, i):
-        """
-        pick i-th buffer from list and display it
-
-        :param i: int
-        :return: None
-        """
-        if len(self.buffers) == 1:
-            # we don't need to display anything
-            # listing is already displayed
-            return
-        else:
-            try:
-                self.display_buffer(self.buffers[i])
-            except IndexError:
-                # i > len
-                self.display_buffer(self.buffers[0])
-
-    @property
-    def current_buffer_index(self):
-        return self.buffers.index(self.current_buffer)
-
-    def remove_current_buffer(self):
-        # don't allow removing main_list
-        if isinstance(self.current_buffer, MainListBuffer):
-            logger.warning("you can't remove main list widget")
-            return
-        self.buffers.remove(self.current_buffer)
-        self.buffer_movement_history.remove(self.current_buffer)
-        self.current_buffer.destroy()
-        self.display_buffer(self.buffer_movement_history[-1], True)
-
-    def unhandled_input(self, key):
-        logger.debug("unhandled input: %r", key)
-        try:
-            if key in ('q', 'Q'):
-                self.worker.shutdown(wait=False)
-                self.ui_worker.shutdown(wait=False)
-                raise urwid.ExitMainLoop()
-            elif key == "ctrl o":
-                self.pick_and_display_buffer(self.current_buffer_index - 1)
-            elif key == "ctrl i":
-                self.pick_and_display_buffer(self.current_buffer_index + 1)
-            elif key == "x":
-                self.remove_current_buffer()
-            elif key == "/":
-                self.prompt("/", search)
-            elif key == "f4":
-                self.mainframe.prompt("filter ", filter)
-            elif key == "n":
-                self.current_buffer.find_next()
-            elif key == "N":
-                self.current_buffer.find_previous()
-            elif key in ["h", "?"]:
-                self.display_help()
-            elif key == "f5":
-                self.display_tree()
-        except NotifyError as ex:
-            self.notify_message(str(ex), level="error")
-            logger.error(repr(ex))
+        self.loop, self.ui = get_app_in_loop(PALLETE)
 
     def run(self):
         self.main_list_buffer = MainListBuffer(self.d, self)
 
+        # TODO move to commands
         # FIXME: this breaks rendering -- focus is not being displayed correctly
         @log_traceback
         def chain_fcs():
@@ -173,6 +33,8 @@ class UI(urwid.MainLoop):
             self.add_and_display_buffer(self.main_list_buffer, redraw=True)
         self.run_quickly_in_bacakground(chain_fcs)
         super().run()
+
+    # TODO: these display methods should be in commands
 
     def display_logs(self, docker_container):
         self.add_and_display_buffer(LogsBuffer(docker_container, self))
