+++ tests/test_concurrency.py
@@ -4,7 +4,7 @@ This suite should test whether sen is capable of running in concurrent high load
 import random
 import threading
 
-from sen.tui.statusbar import UIFrameWidget
+from sen.tui.ui import UIFrameWidget
 from sen.tui.widgets.list.base import VimMovementListBox
 
 import urwid
